Shorten sample files names.
