+++ src/Samples/Basic/Temporal.Samples.DesignProcess/Part1_1_TypicalWorkflow.cs
@@ -10,7 +10,7 @@ using Temporal.Worker.Workflows;
 
 namespace Temporal.Sdk.BasicSamples
 {
-    public class Part01_1_TypicalWorkflow
+    public class Part1_1_TypicalWorkflow
     {
         [Workflow(runMethod: nameof(SayHelloAsync))]
         public class SayHelloWorkflow
@@ -36,7 +36,7 @@ namespace Temporal.Sdk.BasicSamples
         /// <summary>
         /// Parameters to workflow APIs (main method, signal & query parameters) and to activities must implement <see cref="IDataValue" />.
         /// In some specialized cases where it is not possible, the raw (non-deserialized) payload may be accessed
-        /// directly (e.g. <see cref="Part04_1_BasicWorkflowUsage" /> and <see cref="Part04_2_BasicWorkflowUsage_MultipleWorkers" />).
+        /// directly (e.g. <see cref="Part4_1_BasicWorkflowUsage" /> and <see cref="Part4_2_BasicWorkflowUsage_MultipleWorkers" />).
         /// </summary>
         public class SpeechRequest : IDataValue
         {
