+++ src/Samples/Basic/Temporal.Samples.DesignProcess/Part3_1_DynamicWorkflow.cs
@@ -15,7 +15,7 @@ namespace Temporal.Sdk.BasicSamples
 {
     /// <summary>
     /// </summary>
-    public class Part03_1_DynamicWorkflow
+    public class Part3_1_DynamicWorkflow
     {
         /// <summary>A very simple dynamic workflow definition toy model.</summary>
         public class CustomWorkflowDefinition : IDataValue
